+++ satella/coding/transforms/__init__.py
@@ -11,11 +11,23 @@ from .base64 import b64encode
 from .interpol import linear_interpolate
 from .words import hashables_to_int
 
-__all__ = ['stringify', 'split_shuffle_and_join', 'one_tuple',
-           'merge_series', 'pad_to_multiple_of_length', 'clip',
+__all__ = ['stringify', 'split_shuffle_and_join', 'one_tuple', 'none_if_false',
+           'merge_series', 'pad_to_multiple_of_length', 'clip', 'hashables_to_int',
            'jsonify', 'intify', 'percentile', 'b64encode', 'linear_interpolate']
 
 from satella.coding.typing import T, NoArgCallable, Appendable, Number, Predicate
+
+
+def none_if_false(y: tp.Any) -> tp.Optional[tp.Any]:
+    """
+    Return None if y is false, else return y
+
+    :param y: value to check
+    :return: None if y is false, else y
+    """
+    if not y:
+        return None
+    return y
 
 
 def clip(v: Number, minimum: Number, maximum: Number) -> Number: