+++ tests/test_random.py
@@ -1,8 +1,12 @@
 import unittest
-from satella.random import shuffle_together
+from satella.random import shuffle_together, random_binary
 
 
 class TestRandom(unittest.TestCase):
+    def test_random_binary(self):
+        self.assertFalse(random_binary(0))
+        self.assertEqual(len(random_binary(10)), 10)
+
     def test_random(self):
         a = [(1, 2, 3), ('a', 'b', 'c')]
         b = zip(*shuffle_together(*a))