+++ tests/test_coding/test_optionals.py
@@ -1,9 +1,14 @@
 import unittest
 
-from satella.coding.optionals import call_if_nnone, iterate_if_nnone
+from satella.coding.optionals import call_if_nnone, iterate_if_nnone, Optional
 
 
 class TestOptionals(unittest.TestCase):
+
+    def test_optional(self):
+        b = Optional(None)
+        self.assertFalse(b.consumer.cancel().result())
+
     def test_iterate(self):
         b = list(iterate_if_nnone([1, 2, 3]))
         c = list(iterate_if_nnone(None))