+++ satella/coding/decorators/retry_dec.py
@@ -8,7 +8,9 @@ from satella.coding.typing import ExceptionClassType
 def retry(times: tp.Optional[int] = None,
           exc_classes: tp.Union[ExceptionClassType, tp.Tuple[ExceptionClassType, ...]] = Exception,
           on_failure: tp.Callable[[Exception], None] = lambda e: None,
-          swallow_exception: bool = True):
+          swallow_exception: bool = True,
+          call_on_failure: tp.Optional[tp.Callable[[Exception], None]] = None,
+          call_on_success: tp.Optional[tp.Callable[[int], None]] = None):
     """
     A decorator retrying given operation, failing it when an exception shows up.
 
@@ -34,6 +36,10 @@ def retry(times: tp.Optional[int] = None,
         with a single argument, exception instance that was raised last. That exception will
         be swallowed, unless swallow_exception is set to False
     :param swallow_exception: the last exception will be swallowed, unless this is set to False
+    :param call_on_failure: a callable that will be called upon failing to do this, with an
+        exception as it's sole argument. It's result will be discarded.
+    :param call_on_success: a callable that will be called with a single argument: the number
+        of retries that it took to finish the job. It's result will be discarded.
     :return: function result
     """
     def outer(fun):
@@ -43,14 +49,19 @@ def retry(times: tp.Optional[int] = None,
                 iterator = itertools.count()
             else:
                 iterator = range(times)
-            for _ in iterator:
+            for i in iterator:
                 try:
-                    return fun(*args, **kwargs)
+                    y = fun(*args, **kwargs)
+                    if call_on_success is not None:
+                        call_on_success(i)
+                    return y
                 except exc_classes as e:
                     f = e
                     continue
             else:
                 on_failure(f)
+                if call_on_failure is not None:
+                    call_on_failure(f)
                 if not swallow_exception:
                     raise f
         return inner