+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 # v2.14.33
+
+* added `encoding` to `read_lines`