+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 # v2.14.33
 
 * added `encoding` to `read_lines`
+* added `none_if_false`
+* fixed import mishap in `satella.coding.transforms` with `hashables_to_int`