+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 * added `none_if_false`
 * fixed import mishap in `satella.coding.transforms` with `hashables_to_int`
 * fixed `read_in_file` if file does not exist and default is not set
+* add support for `__wrapped__` in `wraps`