+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 # v2.14.34
+
+* added iterate_callable