+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 # v2.14.34
 
 * added iterate_callable
+* added random_binary
+