+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 # v2.14.35
 
+* added AlreadySeen