+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # v2.14.35
 
 * added AlreadySeen
+* fixed a bug with ComparableIntEnum