+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 # v2.14.36
+
+* fixed `AlreadySeen` to support non-hashable objects