+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # v2.14.36
 
 * fixed `AlreadySeen` to support non-hashable objects
+* improved `ComparableIntEnum`