+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 * fixed `AlreadySeen` to support non-hashable objects
 * improved `ComparableIntEnum`
+* fixed a bug in `get_argument` while comparing uncomparable items