+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 # v2.14.37
+
+* added `call_on_failure` and `call_on_success` to `retry`