+++ CHANGELOG.md
@@ -1,3 +1,1 @@
-# v2.14.37
-
-* added `call_on_failure` and `call_on_success` to `retry`
+# v2.14.38